+++ .eslintrc.js
@@ -46,5 +46,6 @@ module.exports = {
     'key-spacing': ['error', { mode: 'strict' }],
     'arrow-spacing': ['error', { before: true, after: true }],
     'react/prop-types': 'off',
+    'react/no-array-index-key': 'off',
   },
 };
